sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentweave-sdk"
version = "0.1.1"
description = "Observability and mesh layer for multi-agent AI systems — track what your agents decided, why they decided it, and how they're connected."
readme = "../../README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Arnab Saha" },
]
keywords = ["agents", "observability", "opentelemetry", "provenance", "tracing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "opentelemetry-api>=1.20",
    "pydantic>=2",
    "typer>=0.9",
    "rich>=13",
]

[project.optional-dependencies]
grpc = ["opentelemetry-exporter-otlp-proto-grpc>=1.20"]
proxy = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.21",
    "build>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
]

[project.scripts]
agentweave = "agentweave.cli:app"

[project.urls]
Homepage = "https://github.com/arnabsaha/agentweave"
Issues = "https://github.com/arnabsaha/agentweave/issues"

[tool.hatch.build.targets.wheel]
packages = ["agentweave"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
